fix form submission
Customer feedback needs a lot of improvement, the design is way too boxy and does not follow the design system.

Excessive use of <br> tags make styling harder, use flexbox instead with a div container. A paragraph normally should not contain a <br> tag, use max-width instead. Refer to the pricing page to see the implementation.

Try to avoid inline style and have all your css written in the external style sheet.

Your overall page design is boring and same, refer to the example given in the google docs - resources